Support using offline test image cache w/ `-m "not needs_internet"`
Disable fetching/updating `.test_images` cache when `-m "not
needs_internet"` is requested.  Instead, either reuse the existing cache
if one is available, or skip the relevant tests.

This preserves the current behavior for the two scenarios: when
`-m "not needs_internet"` is not passed, the cache is normally cloned
and updated.  If it is passed and cache does not exist, the tests are
skipped as before.  However, if one populates the cache first (either
via running the tests or manually cloning it) and then passes `-m "not
needs_internet"`, the tests run against the existing cache.

Fixes #1078