Use typing.IO[bytes] instead of BinaryIO
typing.IO[bytes] is more general (for instance, `io.BytesIO` is an
`IO[bytes]` but not a `BinaryIO`), and we don't need any of the
functionality in `BinaryIO`.

Ideally, we would actually define a custom protocol here (since we
probably don't need the entire file-like API), but that would be a
lot of work.